fractalmesh/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
stripe>=9.0.0
kucoin-python>=1.0.9
python-dotenv>=1.0.1
